Moves DescriptorToDocumentableTransformer to base plugin
