+++ core/src/main/kotlin/transformers/psi/DefaultPsiToDocumentationTranslator.kt
@@ -141,75 +141,3 @@ object DefaultPsiToDocumentationTranslator : PsiToDocumentationTranslator {
         }
     }
 }
-
-enum class JavaClassKindTypes : ClassKind {
-    CLASS,
-    INTERFACE,
-    ENUM_CLASS,
-    ENUM_ENTRY,
-    ANNOTATION_CLASS;
-}
-
-class JavaTypeWrapper : TypeWrapper {
-
-    override val constructorFqName: String?
-    override val constructorNamePathSegments: List<String>
-    override val arguments: List<TypeWrapper>
-    override val dri: DRI?
-    val isPrimitive: Boolean
-
-    constructor(
-        constructorNamePathSegments: List<String>,
-        arguments: List<TypeWrapper>,
-        dri: DRI?,
-        isPrimitiveType: Boolean
-    ) {
-        this.constructorFqName = constructorNamePathSegments.joinToString(".")
-        this.constructorNamePathSegments = constructorNamePathSegments
-        this.arguments = arguments
-        this.dri = dri
-        this.isPrimitive = isPrimitiveType
-    }
-
-    constructor(type: PsiType) {
-        if (type is PsiClassReferenceType) {
-            val resolved = type.resolve()
-            constructorFqName = resolved?.qualifiedName
-            constructorNamePathSegments = resolved?.qualifiedName?.split('.') ?: emptyList()
-            arguments = type.parameters.mapNotNull {
-                if (it is PsiClassReferenceType) JavaTypeWrapper(it) else null
-            }
-            dri = fromPsi(type)
-            this.isPrimitive = false
-        } else if (type is PsiEllipsisType) {
-            constructorFqName = type.canonicalText
-            constructorNamePathSegments = listOf(type.canonicalText) // TODO
-            arguments = emptyList()
-            dri = DRI("java.lang", "Object") // TODO
-            this.isPrimitive = false
-        } else if (type is PsiArrayType) {
-            constructorFqName = type.canonicalText
-            constructorNamePathSegments = listOf(type.canonicalText)
-            arguments = emptyList()
-            dri = (type as? PsiClassReferenceType)?.let { fromPsi(it) } // TODO
-            this.isPrimitive = false
-        } else {
-            type as PsiPrimitiveType
-            constructorFqName = type.name
-            constructorNamePathSegments = type.name.split('.')
-            arguments = emptyList()
-            dri = null
-            this.isPrimitive = true
-        }
-    }
-
-    private fun fromPsi(type: PsiClassReferenceType): DRI {
-        val className = type.className
-        val pkg = type.canonicalText.removeSuffix(className).removeSuffix(".")
-        return DRI(packageName = pkg, classNames = className)
-    }
-
-    override fun toString(): String {
-        return constructorFqName.orEmpty()
-    }
-}
