+++ plugins/base/build.gradle.kts
@@ -6,3 +6,4 @@ publishing {
         }
     }
 }
+
