+++ plugins/kotlin-as-java/build.gradle.kts
@@ -6,3 +6,7 @@ publishing {
         }
     }
 }
+
+dependencies {
+    implementation(project(":plugins:base"))
+}
